Express Routerを使う
